requirements.txt
python-telegram-bot==21.6
anthropic>=0.34.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
